event: rename crossterm module to backend to generalize event interface
